+++ aoj/GRL/GRL1B/main_spfa.cpp
@@ -1,23 +1,21 @@
 //https://onlinejudge.u-aizu.ac.jp/problems/GRL_1_B
-
 #include <bits/stdc++.h>
-
 
 class SPFA {
 private:
     struct edge {
         size_t to;
-        long long cost;
+        int64_t cost;
     };
     size_t N;
     std::vector<std::vector<edge>> edges;
-    long long inf;
+    int64_t inf;
 
 public:
     // Time: O(V)
     SPFA(
         size_t const N = 0,
-        long long const inf = std::numeric_limits<long long>::max() / 2
+        int64_t const inf = std::numeric_limits<int64_t>::max() / 2
     )
         : N(N)
         , edges(N)
@@ -30,12 +28,12 @@ public:
         return N;
     }
     // Time: O(1)
-    long long infinity() const {
+    int64_t infinity() const {
         return inf;
     }
     // from = [0,N), to = [0,N), cost = (-inf,inf)
     // O(1)
-    void add_edge(size_t const from, size_t const to, long long const cost) {
+    void add_edge(size_t const from, size_t const to, int64_t const cost) {
         if (from >= N) throw std::out_of_range("from");
         if (to >= N) throw std::out_of_range("to");
         if (std::abs(cost) >= inf) throw std::out_of_range("cost");
@@ -43,7 +41,7 @@ public:
     }
     // from = [0,N)
     // O (E V)
-    bool solve(size_t const from, std::vector<long long>& D, std::vector<size_t>& P) const {
+    bool solve(size_t const from, std::vector<int64_t>& D, std::vector<size_t>& P) const {
         if (from >= N) throw std::out_of_range("from");
 
         D.assign(N, inf);
@@ -93,7 +91,7 @@ int main() {
         solver.add_edge(u, v, w);
     }
 
-    vector<long long> D;
+    vector<int64_t> D;
     vector<size_t> P;
     bool has_ans = solver.solve(R, D, P);
     if (!has_ans) {
